requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
httplib2
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.22
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0
psycopg2-binary==2.9.1